moved auth logic here to use getInitialProps
